Refactor onboarding to use the installer module
