+++ composer.json
@@ -24,6 +24,13 @@
 	},
 	"repositories": [
 		{
+			"type": "path",
+			"url": "../wp-module-installer",
+			"options": {
+				"symlink": true
+			}
+		},
+		{
 			"type": "composer",
 			"url": "https://newfold-labs.github.io/satis/",
 			"only": [
@@ -38,7 +45,8 @@
 	"require-dev": {
 		"wp-phpunit/wp-phpunit": "^5.9",
 		"yoast/phpunit-polyfills": "^1.0",
-		"newfold-labs/wp-php-standards": "^1.2"
+		"newfold-labs/wp-php-standards": "^1.2",
+		"newfold-labs/wp-module-installer": "@dev"
 	},
 	"scripts": {
 		"lint": [
