+++ includes/Data/Options.php
@@ -33,12 +33,7 @@ final class Options {
 		'auto_update_theme'             => 'auto_update_theme',
 		'permalink_structure'           => 'permalink_structure',
 		'settings_initialized'          => 'settings_initialized',
-		'plugins_init_status'           => 'plugins_init_status',
-		'plugin_install_queue'          => 'plugin_install_queue',
-		'plugin_uninstall_queue'        => 'plugin_uninstall_queue',
 		'flow'                          => 'flow',
-		'theme_init_status'             => 'theme_init_status',
-		'theme_install_queue'           => 'theme_install_queue',
 		'blog_name'                     => 'blogname',
 		'blog_description'              => 'blogdescription',
 		'site_icon'                     => 'site_icon',
