--- includes/Mustache/Templates/aboutPage.mustache
@@ -1,11 +1,0 @@
-<!-- wp:paragraph -->
-<p>Welcome to <span id="nf-1" class="nf-placeholder nf-highlight">organization name</span>. We're dedicated to giving you the very best in <span id="nf-2" class="nf-placeholder nf-highlight">topic/product</span>, with a focus on <span id="nf-3" class="nf-placeholder nf-highlight">brand differentiator 1</span>, <span id="nf-4" class="nf-placeholder nf-highlight">brand differentiator 2</span>, and <span id="nf-5" class="nf-placeholder nf-highlight">brand differentiator 3</span>.</p>
-<!-- /wp:paragraph -->
-
-<!-- wp:paragraph -->
-<p>When we started in <span id="nf-6" class="nf-placeholder nf-highlight">founding date</span> our passion for <span id="nf-7" class="nf-placeholder nf-highlight">product or service type</span> drove us to <span id="nf-8" class="nf-placeholder nf-highlight">an action: create a product, start a community, etc</span>. Now we serve customers all over <span id="nf-9" class="nf-placeholder nf-highlight">your target market area: this could be a city, state, or the whole world</span>, and are thrilled to share our passion with you.</p>
-<!-- /wp:paragraph -->
-
-<!-- wp:paragraph -->
-<p>We hope you enjoy our <span id="nf-10" class="nf-placeholder nf-highlight">products/services</span>. If you have any questions or comments, please contact us <span id="nf-11" class="nf-placeholder nf-highlight">phone number, email, or link to contact form</span>.</p>
-<!-- /wp:paragraph -->
