--- includes/Mustache/Templates/contactPage.mustache
@@ -1,5 +1,0 @@
-<!-- wp:paragraph -->
-<p>Please fill out the form below and we'll be in touch.</p>
-<!-- /wp:paragraph -->
-
-<!-- wp:wpforms/form-selector {"formId":"0"} /-->
