+++ src/OnboardingSPA/utils/api/plugins.js
@@ -1,6 +1,6 @@
 import apiFetch from '@wordpress/api-fetch';
 
-import { onboardingRestURL } from './common';
+import { installerRestURL, onboardingRestURL } from './common';
 import { getQueryParam } from '../index';
 import { resolve } from './resolve';
 import { NFD_PLUGINS_QUERY_PARAM } from '../../../constants';
@@ -14,9 +14,10 @@ export const init = () => {
 		url: onboardingRestURL( 'plugins/initialize' ),
 		method: 'POST',
 		headers: {
-			'X-NFD-ONBOARDING': window.nfdOnboarding.pluginInstallHash,
+			'X-NFD-INSTALLER': window.nfdOnboarding.pluginInstallHash,
 		},
 	} ).catch( ( error ) => {
+		// eslint-disable-next-line no-console
 		console.error( error );
 	} );
 };
@@ -45,7 +46,7 @@ export const setSiteFeatures = async ( pluginInstallHash, data ) => {
 			url: onboardingRestURL( 'plugins/site-features' ),
 			method: 'POST',
 			headers: {
-				'X-NFD-ONBOARDING': pluginInstallHash,
+				'X-NFD-INSTALLER': pluginInstallHash,
 			},
 			data,
 		} )
